+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # _seqr_ Changes
 
 ## dev
+* Add "Validated Name" functional tag (REQUIRES DB MIGRATION)
 
 ## 8/9/24
 * Update directory structure for search backend